move activity items and create reusable show details button
