+++ frontend/pages/hosts/details/cards/Activity/ActivityItems/LockedHostActivityItem/index.ts
