+++ frontend/pages/hosts/details/cards/Activity/ActivityItems/UnlockedHostActivityItem/UnlockedHostActivityItem.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import { IHostActivityItemComponentProps } from "../ActivityConfig";
-import HostActivityItem from "../HostActivityItem";
+import { IHostActivityItemComponentProps } from "../../ActivityConfig";
+import HostActivityItem from "../../HostActivityItem";
 
 const baseClass = "unlocked-host-activity-item";
