+++ frontend/pages/hosts/details/cards/Activity/ShowDetailsButton/_styles.scss
@@ -1,0 +1,5 @@
+.show-details-button {
+  &__show-details-icon {
+    margin-left: $pad-xsmall;
+  }
+}
