+++ frontend/pages/hosts/details/cards/Activity/ShowDetailsButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ShowDetailsButton";
